+++ src/utils/qttrackballwidget.hpp
@@ -109,7 +109,7 @@ public:
     virtual void resizeGL (void)
     {
         camera.setViewport(Eigen::Vector2f ((float)this->width(), (float)this->height()));
-        camera.setPerspectiveMatrix(camera.getFovy(), this->width()/this->height(), 0.1f, 100.0f);
+        camera.setPerspectiveMatrix(camera.getFovy(), (float)this->width()/(float)this->height(), 0.1f, 100.0f);
         light_trackball.setViewport(Eigen::Vector2f ((float)this->width(), (float)this->height()));
         updateGL();
     }
@@ -122,7 +122,7 @@ public:
         Eigen::Vector2i size;
         size << this->width(), this->height();
 
-        camera.setPerspectiveMatrix(60.0, this->width()/this->height(), 0.1f, 100.0f);
+        camera.setPerspectiveMatrix(60.0, (float)this->width()/(float)this->height(), 0.1f, 100.0f);
         camera.setRenderFlag(true);
         camera.setViewport(Eigen::Vector2f ((float)this->width(), (float)this->height()));
